+++ bos/mcu/st/stm32l41x/mcu_stm32l41x_flash.c
@@ -92,11 +92,11 @@ int bMcuFlashLock()
     return retval;
 }
 
-int bMcuFlashErase(uint32_t raddr, uint8_t pages)
+int bMcuFlashErase(uint32_t raddr, uint32_t pages)
 {
     int      retval  = 0;
     int      timeout = 0;
-    uint8_t  i       = 0;
+    uint32_t i       = 0;
     uint32_t tmp;
 
     if (sMcuFlashSize == 0)
